Develop

+++ utils/userUtils.js
@@ -23,11 +23,10 @@ export function setUserLocalStorage(user) {
     }
 
     // fullscreen
-    // If fullscreen is forced, do not overwrite
     if (settings.fullscreen) {
-      if (!localStorage.getItem("fullscreen")) {
+      if (settings.fullscreen != localStorage.getItem("fullscreen") && !localStorage.getItem("fullscreen").includes("force")) {
         localStorage.setItem("fullscreen", settings.fullscreen);
-        store.dispatch(toggleFullscreen(settings.fullscreen));
+        store.dispatch(toggleFullscreen());
       }
     }
